+++ allshoppings-services/src/main/java/mobi/allshoppings/dao/APDAssignationDAO.java
@@ -12,10 +12,10 @@ import mobi.allshoppings.model.APDAssignation;
 public interface APDAssignationDAO extends GenericDAO<APDAssignation> {
 
 	Key createKey(APDAssignation obj) throws ASException;
-	List<APDAssignation> getUsingEntityIdAndEntityKind(String entityId, Integer entityKind) throws ASException;
-	List<APDAssignation> getUsingEntityIdAndEntityKindAndDate(String entityId, Integer entityKind, Date date) throws ASException;
+	List<APDAssignation> getUsingEntityIdAndEntityKind(String entityId, byte entityKind) throws ASException;
+	List<APDAssignation> getUsingEntityIdAndEntityKindAndDate(String entityId, byte entityKind, Date date) throws ASException;
 	List<APDAssignation> getUsingHostnameAndDate(String hostname, Date date) throws ASException;
 	APDAssignation getOneUsingHostnameAndDate(String hostname, Date date) throws ASException;
-	List<String> getEntityIds(Integer entityKind, Date forDate) throws ASException;
+	List<String> getEntityIds(byte entityKind, Date forDate) throws ASException;
 	
 }
