+++ allshoppings-services/src/main/java/mobi/allshoppings/dao/FloorMapDAO.java
@@ -13,7 +13,7 @@ import mobi.allshoppings.tools.Range;
 public interface FloorMapDAO extends GenericDAO<FloorMap> {
 
 	Key createKey(FloorMap obj) throws ASException;
-	List<FloorMap> getUsingStatusAndShoppingId(Integer status, String shoppingId) throws ASException;
-	List<FloorMap> getUsingStatusAndUserAndRange(Integer status, User user, Range range) throws ASException;
+	List<FloorMap> getUsingStatusAndShoppingId(byte status, String shoppingId) throws ASException;
+	List<FloorMap> getUsingStatusAndUserAndRange(byte status, User user, Range range) throws ASException;
 	
 }
